Reorganize showcase so that it can be used as an example
